+++ Includes/Hashers/Hasher.hpp
@@ -6,7 +6,7 @@
 class Hasher
 {
 	public:
-		virtual uint64_t hash(const void* key, size_t len) = 0;
+		virtual uint64_t hash(const void* key, uint64_t len) = 0;
 		virtual uint64_t hash(uint64_t key);
 };
